Updated Payment Info section of the form
Payment Info section of the form. Display payment sections based on
chosen payment option
The"Credit Card" payment option should be selected by default and
result in the display of the #credit-card div, and hide the "Paypal"
and "Bitcoin information.
When a user selects the "PayPal" payment option, display the Paypal
information, and hide the credit card information and the "Bitcoin"
information.
When a user selects the "Bitcoin" payment option, display the Bitcoin
information, and hide the credit card information.